Create STAC Script v1.0
This script is designed for biparental populations, primarily with Recombinant Inbred Lines or Double Haploid populations.
It is meant to facilitate locus calls where genotyping data is sparse.
It examines the markers within a region along a certain chromosome, compares their call data with quality parameters, and attempts to make haplotype calls for the locus. It also graphs out the data in a series of plots that are useful for understanding the structure of the data around this locus, and whether it contains problems (anomalous markers, recombination events, etc.) that complicate matters.